+++ Echo/src/utopia/echo/model/request/OllamaRequest.scala
@@ -4,8 +4,8 @@ import utopia.access.http.Method
 import utopia.access.http.Method.Post
 import utopia.annex.model.request.ApiRequest
 import utopia.disciple.http.request.Body
-import utopia.echo.model.LlmDesignator
 import utopia.echo.model.enumeration.ModelParameter
+import utopia.echo.model.llm.LlmDesignator
 import utopia.flow.generic.casting.ValueConversions._
 import utopia.flow.generic.model.immutable.{Constant, Model, Value}