+++ Echo/src/utopia/echo/model/request/llm/ShowModelRequest.scala
@@ -6,7 +6,7 @@ import utopia.annex.controller.ApiClient
 import utopia.annex.model.request.ApiRequest
 import utopia.annex.model.response.RequestResult
 import utopia.disciple.http.request.Body
-import utopia.echo.model.LlmDesignator
+import utopia.echo.model.llm.LlmDesignator
 import utopia.echo.model.response.llm.ModelShowInfo
 import utopia.flow.generic.model.immutable.{Model, Value}
 import utopia.flow.generic.casting.ValueConversions._