+++ Echo/src/utopia/echo/model/request/llm/StreamedPullRequest.scala
@@ -7,7 +7,7 @@ import utopia.annex.model.request.ApiRequest
 import utopia.annex.model.response.RequestResult
 import utopia.disciple.http.request.Body
 import utopia.echo.controller.parser.StreamedPullResponseParser
-import utopia.echo.model.LlmDesignator
+import utopia.echo.model.llm.LlmDesignator
 import utopia.echo.model.response.llm.StreamedPullStatus
 import utopia.flow.generic.casting.ValueConversions._
 import utopia.flow.generic.model.immutable.{Model, Value}